Rename useResponsiveSize to useResponsiveCanvasSize and improve the hook, add input trimming to avoid sending empty spaces inside WelcomeModal.tsx
